requirements.txt
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
psycopg2-binary==2.9.9
requests==2.31.0
django-filter==23.3
drf-yasg==1.21.7
fcm-django==1.0.12
django-extensions==3.2.3